UFC_ASK_API/verify_token.py
from fastapi import Depends, HTTPException, status
from jose import JWTError, jwt
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import os

#mesma chave secreta usada no Node.js
SECRET_KEY = os.getenv("JWT_SECRET") 
ALGORITHM = os.getenv("ALG")  # ou o mesmo usado na aplicação Node.js

security = HTTPBearer()

def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    token = credentials.credentials
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload  
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Invalid or expired token",
        )
